Added a script for Ubuntu
The regex in the original script did not work in Ubuntu so a new script
was created. This script was tested in Ubuntu 12.10 but should work in
earlier versions of Ubuntu as well as most Linux flavors. Eventually I
will test this new script in Mac as well and attempt to come up with a
unified script for both Mac and Linux.